+++ controllers/front/ajax.php
@@ -15,10 +15,11 @@
 if (!defined('_PS_VERSION_')) {
     exit;
 }
+use PrestaShop\Module\Doofinder\Api\DoofinderApi;
+use PrestaShop\Module\Doofinder\Installer\DoofinderInstallation;
+use PrestaShop\Module\Doofinder\Utils\DfTools;
 
-use PrestaShop\Module\Doofinder\Entity\DfTools;
-use PrestaShop\Module\Doofinder\Entity\DoofinderApi;
-use PrestaShop\Module\Doofinder\Entity\DoofinderInstallation;
+
 
 class DoofinderAjaxModuleFrontController extends ModuleFrontController
 {
