+++ src/Core/DoofinderResults.php
@@ -12,7 +12,7 @@
  * - getResults: get an array with the results
  */
 
-namespace PrestaShop\Module\Doofinder\Entity;
+namespace PrestaShop\Module\Doofinder\Core;
 
 if (!defined('_PS_VERSION_')) {
     exit;
