+++ upgrade/upgrade-4.12.0.php
@@ -24,7 +24,7 @@
  * International Registered Trademark & Property of PrestaShop SA
  */
 
-use PrestaShop\Module\Doofinder\Entity\DoofinderConfig;
+use PrestaShop\Module\Doofinder\Configuration\DoofinderConfig;
 
 if (!defined('_PS_VERSION_')) {
     exit;
